requirements.txt
# flask>=3.0.0
streamlit>=1.35.0
mediapipe>=0.10.0
opencv-python-headless>=4.9.0.80
numpy>=1.24.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pillow>=10.0.0

